Split the payment test-button tests under the file limit
The suite reached 443 lines, above the 400-line limit `AGENTS.md` sets for
a test file. It is now a folder: `page.ts` installs the happy-dom page and
takes it down, `button.test.ts` covers what the button does whichever
provider it belongs to, and `answers.test.ts` covers the words each
provider's answer turns into.

The mirror path still holds, because a directory at the source's path
answers the mutation gate as well as a file does. The kill rate is
unchanged at 100%.
